show unit info on panel
